+++ src/ex/ex.cpp
@@ -1,7 +1,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 // Name:      ex.cpp
 // Purpose:   Implementation of class wex::ex
-//            http://pubs.opengroup.org/onlinepubs/9699919799/utilities/ex.html
+//            https://pubs.opengroup.org/onlinepubs/9799919799/utilities/ex.html
 // Author:    Anton van Wezenbeek
 // Copyright: (c) 2012-2024 Anton van Wezenbeek
 ////////////////////////////////////////////////////////////////////////////////